Warning hunt.

+++ avs_core/audio/avs-soundtouch.cpp
@@ -321,17 +321,17 @@ AVSValue __cdecl AVSsoundtouch::Create(AVSValue args, void*, IScriptEnvironment*
 	try {	// HIDE DAMN SEH COMPILER BUG!!!
   if (args[0].AsClip()->GetVideoInfo().AudioChannels() == 2) {
     return new AVSStereoSoundTouch(args[0].AsClip(), 
-      args[1].AsFloat(100.0), 
-      args[2].AsFloat(100.0), 
-      args[3].AsFloat(100.0), 
-	  &args[4],
+      (float)args[1].AsFloat(100.0), 
+      (float)args[2].AsFloat(100.0), 
+      (float)args[3].AsFloat(100.0), 
+	    &args[4],
       env);
   }
   return new AVSsoundtouch(args[0].AsClip(), 
-    args[1].AsFloat(100.0), 
-    args[2].AsFloat(100.0), 
-    args[3].AsFloat(100.0), 
-	&args[4],
+    (float)args[1].AsFloat(100.0), 
+    (float)args[2].AsFloat(100.0), 
+    (float)args[3].AsFloat(100.0), 
+	  &args[4],
     env);
 
 	}
