+++ avs_core/filters/fps.cpp
@@ -276,12 +276,12 @@ AVSValue __cdecl ContinuedCreate(AVSValue args, void* key, IScriptEnvironment* e
     reduce_frac(num, den, (unsigned)args[2].AsInt(1001));
   } else { // float[, limit] form
     if (args[2].IsInt()) {
-      if (float_to_frac(args[0].AsFloat(), num, den)) {
+      if (float_to_frac((float)args[0].AsFloat(), num, den)) {
         env->ThrowError("ContinuedFraction: Float value out of range for rational pair.\n");
       }
       reduce_frac(num, den, (unsigned)args[2].AsInt());
     } else {
-      if (reduce_float(args[0].AsFloat(), num, den)) {
+      if (reduce_float((float)args[0].AsFloat(), num, den)) {
         env->ThrowError("ContinuedFraction: Float value out of range for rational pair.\n");
       }
     }
@@ -455,7 +455,7 @@ AVSValue __cdecl AssumeFPS::CreateFloat(AVSValue args, void*, IScriptEnvironment
 	unsigned num, den;
 
 	try {	// HIDE DAMN SEH COMPILER BUG!!!
-	  FloatToFPS("AssumeFPS", args[1].AsFloat(), num, den, env);
+	  FloatToFPS("AssumeFPS", (float)args[1].AsFloat(), num, den, env);
 	  return new AssumeFPS(args[0].AsClip(), num, den, args[2].AsBool(false), env);
 	}
 	catch (...) { throw; }
@@ -546,7 +546,7 @@ AVSValue __cdecl ChangeFPS::CreateFloat(AVSValue args, void*, IScriptEnvironment
 	unsigned num, den;
 
 	try {	// HIDE DAMN SEH COMPILER BUG!!!
-	  FloatToFPS("ChangeFPS", args[1].AsFloat(), num, den, env);
+	  FloatToFPS("ChangeFPS", (float)args[1].AsFloat(), num, den, env);
 	  return new ChangeFPS(args[0].AsClip(), num, den, args[2].AsBool(true), env);
 	}
 	catch (...) { throw; }
@@ -760,7 +760,7 @@ AVSValue __cdecl ConvertFPS::CreateFloat(AVSValue args, void*, IScriptEnvironmen
 	unsigned num, den;
 
 	try {	// HIDE DAMN SEH COMPILER BUG!!!
-	  FloatToFPS("ConvertFPS", args[1].AsFloat(), num, den, env);
+	  FloatToFPS("ConvertFPS", (float)args[1].AsFloat(), num, den, env);
 	  return new ConvertFPS( args[0].AsClip(), num, den, args[2].AsInt(-1), args[3].AsInt(0), env );
 	}
 	catch (...) { throw; }
